Module reth_node_api::engine

Traits, validation methods, and helper types used to abstract over engine types.

Notably contains the EngineTypes trait and implementations for ethereum mainnet types. This contains the EngineTypes trait and implementations for ethereum mainnet types.

The EngineTypes trait can be implemented to configure the engine to work with custom types, as long as those types implement certain traits.

Custom payload attributes can be supported by implementing two main traits:

PayloadAttributes can be implemented for payload attributes types that are used as arguments to the engine_forkchoiceUpdated method. This type should be used to define and spawn payload jobs.

PayloadBuilderAttributes can be implemented for payload attributes types that describe running payload jobs.

Once traits are implemented and custom types are defined, the EngineTypes trait can be implemented:

/// A custom payload attributes type.
#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
pub struct CustomPayloadAttributes {
    /// An inner payload type
    #[serde(flatten)]
    pub inner: EthPayloadAttributes,
    /// A custom field
    pub custom: u64,
}

/// Custom error type used in payload attributes validation
#[derive(Debug, Error)]
pub enum CustomError {
   #[error("Custom field is not zero")]
   CustomFieldIsNotZero,
}

impl PayloadAttributes for CustomPayloadAttributes {
    fn timestamp(&self) -> u64 {
        self.inner.timestamp()
    }

    fn withdrawals(&self) -> Option<&Vec<Withdrawal>> {
        self.inner.withdrawals()
    }

    fn parent_beacon_block_root(&self) -> Option<B256> {
        self.inner.parent_beacon_block_root()
    }

    fn ensure_well_formed_attributes(
        &self,
        chain_spec: &ChainSpec,
        version: EngineApiMessageVersion,
    ) -> Result<(), AttributesValidationError> {
        validate_version_specific_fields(chain_spec, version, self.into())?;

        // custom validation logic - ensure that the custom field is not zero
        if self.custom == 0 {
            return Err(AttributesValidationError::invalid_params(
                CustomError::CustomFieldIsNotZero,
            ))
        }

        Ok(())
    }
}

/// Newtype around the payload builder attributes type
#[derive(Clone, Debug, PartialEq, Eq)]
pub struct CustomPayloadBuilderAttributes(EthPayloadBuilderAttributes);

impl PayloadBuilderAttributes for CustomPayloadBuilderAttributes {
    type RpcPayloadAttributes = CustomPayloadAttributes;
    type Error = Infallible;

    fn try_new(parent: B256, attributes: CustomPayloadAttributes) -> Result<Self, Infallible> {
        Ok(Self(EthPayloadBuilderAttributes::new(parent, attributes.inner)))
    }

    fn parent(&self) -> B256 {
        self.0.parent
    }

    fn payload_id(&self) -> PayloadId {
        self.0.id
    }

    fn timestamp(&self) -> u64 {
        self.0.timestamp
    }

    fn parent_beacon_block_root(&self) -> Option<B256> {
        self.0.parent_beacon_block_root
    }

    fn suggested_fee_recipient(&self) -> Address {
        self.0.suggested_fee_recipient
    }

    fn prev_randao(&self) -> B256 {
        self.0.prev_randao
    }

    fn withdrawals(&self) -> &Vec<reth_primitives::Withdrawal> {
        &self.0.withdrawals
    }
}

/// Custom engine types - uses a custom payload attributes RPC type, but uses the default
/// payload builder attributes type.
#[derive(Clone, Debug, Default)]
#[non_exhaustive]
pub struct CustomEngineTypes;

impl EngineTypes for CustomEngineTypes {
   type PayloadAttributes = CustomPayloadAttributes;
   type PayloadBuilderAttributes = CustomPayloadBuilderAttributes;
   type BuiltPayload = EthBuiltPayload;

   fn validate_version_specific_fields(
       chain_spec: &ChainSpec,
       version: EngineApiMessageVersion,
       payload_or_attrs: PayloadOrAttributes<'_, CustomPayloadAttributes>,
   ) -> Result<(), AttributesValidationError> {
       validate_version_specific_fields(chain_spec, version, payload_or_attrs)
   }
}

Re-exports

pub use traits::BuiltPayload;

pub use traits::PayloadAttributes;

pub use traits::PayloadBuilderAttributes;

pub use error::AttributesValidationError;

pub use payload::PayloadOrAttributes;

Modules

error

Contains error types used in the traits defined in this crate. Defines a payload validation error type

payload

Contains types used in implementations of the PayloadAttributes trait.

traits

Contains traits to abstract over payload attributes types and default implementations of the PayloadAttributes trait for ethereum mainnet and optimism types.

Enums

EngineApiMessageVersion

The version of Engine API message.

Traits

EngineTypes

The types that are used by the engine.

Functions

validate_parent_beacon_block_root_presence

Validate the presence of the parentBeaconBlockRoot field according to the payload timestamp.

validate_payload_timestamp

Validates the timestamp depending on the version called:

validate_version_specific_fields

Validates the presence or exclusion of fork-specific fields based on the payload attributes and the message version.

validate_withdrawals_presence

Validates the presence of the withdrawals field according to the payload timestamp. After Shanghai, withdrawals field must be Some. Before Shanghai, withdrawals field must be None;