Trait reth_node_api::engine::traits::BuiltPayload

pub trait BuiltPayload: Send + Sync + Debug {
    // Required methods
    fn block(&self) -> &SealedBlock;
    fn fees(&self) -> U256;
    fn into_v1_payload(self) -> ExecutionPayloadV1;
    fn into_v2_payload(self) -> ExecutionPayloadEnvelopeV2;
    fn into_v3_payload(self) -> ExecutionPayloadEnvelopeV3;
}

Represents a built payload type that contains a built SealedBlock and can be converted into engine API execution payloads.

Required Methods

fn block(&self) -> &SealedBlock

Returns the built block (sealed)

fn fees(&self) -> U256

Returns the fees collected for the built block

fn into_v1_payload(self) -> ExecutionPayloadV1

Converts the type into the response expected by engine_getPayloadV1

fn into_v2_payload(self) -> ExecutionPayloadEnvelopeV2

Converts the type into the response expected by engine_getPayloadV2

fn into_v3_payload(self) -> ExecutionPayloadEnvelopeV3

Converts the type into the response expected by engine_getPayloadV2