Trait reth_node_api::evm::ConfigureEvmEnv

pub trait ConfigureEvmEnv: Send + Sync + Unpin + Clone {
    type TxMeta;

    // Required methods
    fn fill_tx_env<T>(
        tx_env: &mut TxEnv,
        transaction: T,
        sender: Address,
        meta: Self::TxMeta
    )
       where T: AsRef<Transaction>;
    fn fill_cfg_env(
        cfg_env: &mut CfgEnv,
        chain_spec: &ChainSpec,
        header: &Header,
        total_difficulty: U256
    );

    // Provided method
    fn fill_cfg_and_block_env(
        cfg: &mut CfgEnv,
        block_env: &mut BlockEnv,
        chain_spec: &ChainSpec,
        header: &Header,
        total_difficulty: U256
    ) { ... }
}

This represents the set of methods used to configure the EVM before execution.

Required Associated Types

type TxMeta

The type of the transaction metadata.

Required Methods

fn fill_tx_env<T>( tx_env: &mut TxEnv, transaction: T, sender: Address, meta: Self::TxMeta )
where T: AsRef<Transaction>,

Fill transaction environment from a Transaction and the given sender address.

fn fill_cfg_env( cfg_env: &mut CfgEnv, chain_spec: &ChainSpec, header: &Header, total_difficulty: U256 )

Fill CfgEnv fields according to the chain spec and given header

Provided Methods

fn fill_cfg_and_block_env( cfg: &mut CfgEnv, block_env: &mut BlockEnv, chain_spec: &ChainSpec, header: &Header, total_difficulty: U256 )

Convenience function to call both fill_cfg_env and fill_block_env.

Object Safety

This trait is not object safe.