Module reth_node_api::engine

Traits, validation methods, and helper types used to abstract over engine types.

Notably contains the EngineTypes trait and implementations for ethereum mainnet types. This contains the EngineTypes trait and implementations for ethereum mainnet types.

Re-exports

pub use traits::BuiltPayload;

pub use traits::PayloadAttributes;

pub use traits::PayloadBuilderAttributes;

pub use error::AttributesValidationError;

pub use payload::PayloadOrAttributes;

Modules

error

Contains error types used in the traits defined in this crate. Defines a payload validation error type

payload

Contains types used in implementations of the PayloadAttributes trait.

traits

Contains traits to abstract over payload attributes types and default implementations of the PayloadAttributes trait for ethereum mainnet and optimism types.

Enums

EngineApiMessageVersion

The version of Engine API message.

Traits

EngineTypes

The types that are used by the engine.

Functions

validate_parent_beacon_block_root_presence

Validate the presence of the parentBeaconBlockRoot field according to the payload timestamp.

validate_payload_timestamp

Validates the timestamp depending on the version called:

validate_version_specific_fields

Validates the presence or exclusion of fork-specific fields based on the ethereum payload attributes and the message version.

validate_withdrawals_presence

Validates the presence of the withdrawals field according to the payload timestamp. After Shanghai, withdrawals field must be Some. Before Shanghai, withdrawals field must be None;